mu: thread: sort only root containers according to the leader
Today when we query a find cmd with the `--threads` option, all the
threads/childs are sorted according to their leaders based on the
sortfield.

This patch change the way how to sort the threads.
The threads are still sorted according to the leader of each thread
but all the childs of each thread are now sorted based on the
sortfield only.

The previous implementation worked fine but for some cases the results
can be unexpected.
For example when we subscribe to kernel mailing-list, the patchs
received by email are numbered.

Example with random thread:
[PATCH 0/4] drm/panel: jh057n0090: Add regulators and drop magic value in init
  ┣━▶[PATCH 1/4] MAINTAINERS: Add Purism mail alias as reviewer for their devkit's panel
  ┣━▶[PATCH 2/4] drm/panel: jh057n0090: Don't use magic constant
  ┣━▶[PATCH 3/4] dt-bindings: display/panel: jh057n0090: Document power supply properties
  ┗━▶[PATCH 4/4] drm/panel: jh057n0090: Add regulator support

If someone reply to one these emails in the middle, this email become
the leader and the thread is displayed like this:

[PATCH 0/4] drm/panel: jh057n0090: Add regulators and drop magic value in init
  ┣━▶[PATCH 2/4] drm/panel: jh057n0090: Don't use magic constant
  ┃  ┗━▶ Re: [PATCH 2/4] drm/panel: jh057n0090: Don't use magic constant
  ┣━▶[PATCH 3/4] dt-bindings: display/panel: jh057n0090: Document power supply properties
  ┣━▶[PATCH 1/4] MAINTAINERS: Add Purism mail alias as reviewer for their devkit's panel
  ┗━▶[PATCH 4/4] drm/panel: jh057n0090: Add regulator support

So if we have multiple replies for different child of the thread, the
initial order can be completely disaligned, especially when we want to
sort according to the date (which is most of the case).

This patch handle differently the sorting.
The leader of a root container (a thread without parent) is
interesting only we want to sort all the root containers which are at
level 0.
Indeed when we sort by date, the most recent thread will be on
top/bottom (depends on sorting direction).
However all the childs of a thread (level > 0) should be sorted based
on the sortfield only not the leader.

With the example above, this lead to the kind of output:
[PATCH 0/4] drm/panel: jh057n0090: Add regulators and drop magic value in init
  ┣━▶[PATCH 1/4] MAINTAINERS: Add Purism mail alias as reviewer for their devkit's panel
  ┣━▶[PATCH 2/4] drm/panel: jh057n0090: Don't use magic constant
  ┃  ┗━▶ Re: [PATCH 2/4] drm/panel: jh057n0090: Don't use magic constant
  ┣━▶[PATCH 3/4] dt-bindings: display/panel: jh057n0090: Document power supply properties
  ┗━▶[PATCH 4/4] drm/panel: jh057n0090: Add regulator support

Another example with testdir3 from the test suite:

Before patch:
$ mu find --nocolor --threads --reverse maildir:'/tree'
  `*> Fri 24 Jun 2011 01:20:00 PM CEST testfrom@example.com Re: child 4.1
  |*> Fri 24 Jun 2011 01:10:00 PM CEST testfrom@example.com Re: child 4.0
Wed 22 Jun 2011 06:33:00 PM CEST testfrom@example.com Re: child 3.0.0.0
Tue 21 Jun 2011 03:00:00 PM CEST testfrom@example.com root2
  `-> Tue 21 Jun 2011 05:02:00 PM CEST testfrom@example.com Re: child 2.0.0
Tue 21 Jun 2011 02:00:00 PM CEST testfrom@example.com root1
Tue 21 Jun 2011 01:00:00 PM CEST testfrom@example.com root0
  `-> Tue 21 Jun 2011 01:20:00 PM CEST testfrom@example.com Re: child 0.1
    `-> Tue 21 Jun 2011 01:22:00 PM CEST testfrom@example.com Re: child 0.1.0
  |-> Tue 21 Jun 2011 01:10:00 PM CEST testfrom@example.com Re: child 0.0

After patch:
$ mu find --nocolor --threads --reverse maildir:'/tree'
  `*> Fri 24 Jun 2011 01:10:00 PM CEST testfrom@example.com Re: child 4.0
  |*> Fri 24 Jun 2011 01:20:00 PM CEST testfrom@example.com Re: child 4.1
Wed 22 Jun 2011 06:33:00 PM CEST testfrom@example.com Re: child 3.0.0.0
Tue 21 Jun 2011 03:00:00 PM CEST testfrom@example.com root2
  `-> Tue 21 Jun 2011 05:02:00 PM CEST testfrom@example.com Re: child 2.0.0
Tue 21 Jun 2011 02:00:00 PM CEST testfrom@example.com root1
Tue 21 Jun 2011 01:00:00 PM CEST testfrom@example.com root0
  `-> Tue 21 Jun 2011 01:10:00 PM CEST testfrom@example.com Re: child 0.0
  |-> Tue 21 Jun 2011 01:20:00 PM CEST testfrom@example.com Re: child 0.1
    `-> Tue 21 Jun 2011 01:22:00 PM CEST testfrom@example.com Re: child 0.1.0

`test-mu-threads` from the test suite has also been updated since the
order is now different.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>